feat!: CLI cleanup - output format standardization and command restructuring
BREAKING CHANGES:
- Replace --json/-j with --output-format/-f enum (Text, Json) across all commands (#73)
- Move schema and users commands to data group (#74)
  - ppds schema generate → ppds data schema
  - ppds users generate → ppds data users
- Remove ppds schema list command (#74)

Other changes:
- Remove Newtonsoft.Json dependency from PPDS.Dataverse, use System.Text.Json (#72)
- Fix .gitignore data/ pattern to only match repo root

Closes #72, #73, #74
