+++ src/PPDS.Cli/Infrastructure/ServiceFactory.cs
@@ -1,4 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
+using PPDS.Cli.Commands;
 using PPDS.Migration.Analysis;
 using PPDS.Migration.Formats;
 using PPDS.Migration.Progress;
@@ -11,14 +12,14 @@ namespace PPDS.Cli.Infrastructure;
 public static class ServiceFactory
 {
     /// <summary>
-    /// Creates a progress reporter based on the output mode.
+    /// Creates a progress reporter based on the output format.
     /// </summary>
-    /// <param name="useJson">Whether to output JSON format.</param>
+    /// <param name="outputFormat">The output format.</param>
     /// <param name="operationName">The operation name for completion messages (e.g., "Export", "Import").</param>
     /// <returns>An appropriate progress reporter.</returns>
-    public static IProgressReporter CreateProgressReporter(bool useJson, string operationName = "Operation")
+    public static IProgressReporter CreateProgressReporter(OutputFormat outputFormat, string operationName = "Operation")
     {
-        IProgressReporter reporter = useJson
+        IProgressReporter reporter = outputFormat == OutputFormat.Json
             ? new JsonProgressReporter(Console.Out)
             : new ConsoleProgressReporter();
