+++ src/PPDS.Dataverse/PPDS.Dataverse.csproj
@@ -52,7 +52,6 @@
     <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="10.0.1" />
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
     <PackageReference Include="MinVer" Version="6.1.0" PrivateAssets="all" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.*" />
     <!-- Override vulnerable transitive dependency from Dataverse.Client (GHSA-555c-2p6r-68mm) -->
     <PackageReference Include="System.Security.Cryptography.Pkcs" Version="10.0.1" />
   </ItemGroup>
