+++ tests/PPDS.Cli.Tests/Commands/Plugins/CleanCommandTests.cs
@@ -81,7 +81,7 @@ public class CleanCommandTests : IDisposable
     [Fact]
     public void Create_HasOptionalJsonOption()
     {
-        var option = _command.Options.FirstOrDefault(o => o.Name == "--json");
+        var option = _command.Options.FirstOrDefault(o => o.Name == "--output-format");
         Assert.NotNull(option);
         Assert.False(option.Required);
     }
@@ -135,7 +135,7 @@ public class CleanCommandTests : IDisposable
     [Fact]
     public void Parse_WithOptionalJson_Succeeds()
     {
-        var result = _command.Parse($"-c \"{_tempConfigFile}\" --json");
+        var result = _command.Parse($"-c \"{_tempConfigFile}\" --output-format Json");
         Assert.Empty(result.Errors);
     }
 
@@ -147,7 +147,7 @@ public class CleanCommandTests : IDisposable
             "--profile dev " +
             "--environment https://org.crm.dynamics.com " +
             "--what-if " +
-            "--json");
+            "--output-format Json");
         Assert.Empty(result.Errors);
     }
